+++ libblankenhain/src/Filter.cpp
@@ -45,7 +45,7 @@ void Filter<T>::setNotch(double frequency, double Q)
 template <typename T>
 void Filter<T>::setBell(double frequency, double Q, double gainInDb)
 {
-	double A = BhMath::pow(10, gainInDb / 40.);
+	double A = static_cast<double>(BhMath::pow(10, static_cast<float>(gainInDb / 40.)));
 	recomputeCoefficients(frequency, Q * A, 1);
 	m0 = T(1);
 	m1 = T(k * T(A * A - 1));
@@ -55,8 +55,8 @@ void Filter<T>::setBell(double frequency, double Q, double gainInDb)
 template <typename T>
 void Filter<T>::setLowShelf(double frequency, double Q, double gainInDb)
 {
-	double A = BhMath::pow(10, gainInDb / 40.);
-	recomputeCoefficients(frequency, Q, 1. / BhMath::sqrt(A));
+	double A = static_cast<double>(BhMath::pow(10.f, static_cast<float>(gainInDb) / 40.f));
+	recomputeCoefficients(frequency, Q, 1. / BhMath::sqrt(static_cast<float>(A)));
 	m0 = T(1);
 	m1 = k * T(A - 1);
 	m2 = T(A * A - 1);
@@ -65,8 +65,8 @@ void Filter<T>::setLowShelf(double frequency, double Q, double gainInDb)
 template <typename T>
 void Filter<T>::setHighShelf(double frequency, double Q, double gainInDb)
 {
-	double A = BhMath::pow(10, gainInDb / 40.);
-	recomputeCoefficients(frequency, Q, BhMath::sqrt(A));
+	double A = static_cast<double>(BhMath::pow(10.f, static_cast<float>(gainInDb) / 40.f));
+	recomputeCoefficients(frequency, Q, BhMath::sqrt(static_cast<float>(A)));
 	m0 = T(A * A);
 	m1 = k * T((1 - A) * A);
 	m2 = T(1 - A * A);
@@ -104,7 +104,7 @@ T _vectorcall Filter<T>::tick(const T& v0)
 template <typename T>
 void Filter<T>::recomputeCoefficients(double frequency, double Q, double gFactor)
 {
-	double g_ = BhMath::tan(constants::pi * frequency / constants::sampleRate) * gFactor;
+	double g_ = static_cast<double>(BhMath::tan(constants::pi * static_cast<float>(frequency) / constants::sampleRate)) * gFactor;
 	double k_ = 1. / Q;
 	double a1_ = 1. / (1. + g_ * (g_ + k_));
 	double a2_ = g_ * a1_;
